src/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
asyncpg==0.29.0
alembic==1.13.1
PyJWT==2.8.0
decode==2025.2.0
dotenv==0.9.9
